Time: 0 ms (100%), Space: 8.8 MB (83.16%) - LeetHub

+++ 1404-number-of-steps-to-reduce-a-number-in-binary-representation-to-one/1404-number-of-steps-to-reduce-a-number-in-binary-representation-to-one.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int numSteps(string s) {
-        int n = (int)s.size();
+        int n = s.size();
         int carry = 0;
         int steps = 0;
